Cast conversions from NSUInteger to unsigned long
Errors are surfaced when attempting command-line builds, which
apparently use stricter error reporting.

+++ Fox/Private/Generators/FOXSuchThatGenerator.m
@@ -34,7 +34,8 @@
     }
 
     [NSException raise:@"FOXGenerationFailed"
-                format:@"FOXSuchThat generator could failed to satisfy predicate after %lu times", self.maxTries];
+                format:@"FOXSuchThat generator could failed to satisfy predicate after %lu times",
+                       (unsigned long)self.maxTries];
     return nil;
 }
